requirements.txt
fastapi>=0.95.0
uvicorn[standard]
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
requests
pydantic
jira
pytest

# Model and runtime dependencies for summarization
# Pin transformer and torch to versions known to work together in this project.
# If you have a CUDA GPU, install the matching torch wheel per https://pytorch.org
sentencepiece
# Optional (Mistral / quantized GPU): uncomment if needed
# accelerate
# bitsandbytes
# safetensors
transformers>=4.36.0
torch>=2.0.1
spacy